requirements.txt
Pillow
wikipedia
pandas
spacy
fuzzywuzzy
animation
unidecode
